+++ src/inputs.ts
@@ -18,7 +18,7 @@ const STRING = "string";
 const TRUE = "true";
 
 const CONVENTIONAL_COMMIT_TITLE = "chore: optimize {{optimizedCount}} SVG(s)";
-const DEFAULT_COMMIT_DESCRIPTION = "Optimized SVG(s):\n{{filesList}}";
+const DEFAULT_COMMIT_BODY = "Optimized SVG(s):\n{{filesList}}";
 const DEFAULT_COMMIT_TITLE = "Optimize {{optimizedCount}} SVG(s) with SVGO";
 const DEFAULT_COMMENT = "SVG(s) automatically optimized using [SVGO](https://github.com/svg/svgo) :sparkles:\n\n{{filesTable}}";
 
@@ -28,7 +28,7 @@ export type RawActionConfig = {
   readonly commit?: {
     readonly conventional?: boolean;
     readonly title?: string;
-    readonly description?: string;
+    readonly body?: string;
   };
   readonly "dry-run"?: boolean;
   readonly ignore?: string;
@@ -47,7 +47,7 @@ export function getRepoToken(): string {
 export class ActionConfig {
 
   public readonly comment: string;
-  public readonly commitDescription: string;
+  public readonly commitBody: string;
   public readonly commitTitle: string;
   public readonly enableComments: boolean;
   public readonly ignoreGlob: string;
@@ -56,7 +56,7 @@ export class ActionConfig {
 
   constructor(config: RawActionConfig = { }) {
     this.comment = ActionConfig.getCommentValue(config);
-    this.commitDescription = ActionConfig.getCommitDescription(config);
+    this.commitBody = ActionConfig.getCommitBody(config);
     this.commitTitle = ActionConfig.getCommitTitle(config);
     this.enableComments = ActionConfig.getEnableCommentsValue(config);
     this.ignoreGlob = ActionConfig.getIgnoreGlob(config);
@@ -78,9 +78,9 @@ export class ActionConfig {
     }
   }
 
-  private static getCommitDescription(config: RawActionConfig): string {
-    return (config.commit?.description !== undefined) ?
-      config.commit.description : DEFAULT_COMMIT_DESCRIPTION;
+  private static getCommitBody(config: RawActionConfig): string {
+    return (config.commit?.body !== undefined) ?
+      config.commit.body : DEFAULT_COMMIT_BODY;
   }
 
   private static getCommitTitle(config: RawActionConfig): string {